+++ RELEASE_NOTES.md
@@ -2,7 +2,8 @@ Release 1.0.9
 
 - Add detection for 6 GHz out-of-band as a supported alternative operating class
 - Add detection for 6 GHz in-band for band capabilities found when the client associations in 6 GHz
-- Enhance `--list_interface` output formatting
+- Improve `--list_interface` output
+- Add `--debug` option as shorthand for `--logging debug`
 
 Release 1.0.8-2